SOHO Router, Switch, and Router Hardening
You can't understand how to harden a device until you understand how it communicates and operates.  The first portion of Security+ is developing that baseline knowledge.  Near the end of the course we apply hardening principles to devices in a Cisco Packet Capture network.  While the specific Cisco commands are not necessary for the student to know during the exam, the process and principles are.  Walking through this lab, even if it is with the key, cements the concepts of device hardening.